backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain-core
python-dotenv
pydantic